Compact search indexes and release them when searching goes idle
Postings hold slot numbers as UInt32, the width they already had on disk,
in both the full and short-query indexes. IndexedItem drops the five fields
search never reads (content hash, creation date, use count, size, storage
ref); the full-index disk cache moves to v6 and the existing stale-file
sweep removes v5 files.

After 60 s without a search the engine releases the recent-items cache, the
sorted-match cache, cached statements, and SQLite heap memory on its
connection, and releases the full index, first persisting it when its disk
cache is older than memory; malloc pressure relief then returns freed pages.
The next session loads the index through the existing disk-cache or
database path. ClipboardService watches memory pressure: a warning trims
the same way without persisting, a critical event also drops the
short-query index and releases the write connection's SQLite memory.

Snapshot release bench (fresh perf-db copy): cmd p95 0.51 ms, prepared cm
p95 4.75 ms, cold cm 47.8 ms. Warm load (v6 cache): 50.2 ms, peak RSS
103.1 MB, peak footprint 66.5 MB, cache payload 25.4 MB.

+++ ScopyTests/FullIndexDiskCacheHardeningTests.swift
@@ -221,7 +221,7 @@ final class FullIndexDiskCacheHardeningTests: XCTestCase {
         XCTAssertEqual(asciiPostings.count, 128)
 
         let iIndex = Int(Character("i").asciiValue ?? 0)
-        asciiPostings[iIndex].append(itemsCount)
+        asciiPostings[iIndex].append(UInt32(itemsCount))
 
         let corruptedData = SearchIndexDiskCache.debugEncodeFullPayload(payload, asciiCharPostings: asciiPostings)
         try corruptedData.write(to: cacheURL, options: [.atomic])
